+++ .eslintrc.json
@@ -6,5 +6,6 @@
     "plugins": [
         "n",
         "putout"
-    ]
+    ],
+    "ignorePatterns": ["*/fixture/**/*.js"]
 }
