+++ lib/index.js
@@ -9,7 +9,6 @@ const plugin = require('./plugin');
 const test = async () => {
     const source = await readFile(join(__dirname, '../test/fixture/source.js'), 'utf-8');
     
-    console.log(source);
     const {places} = putout(source, {
         fix: false,
         plugins: [
