+++ lib/plugin/+react-import/index.js
@@ -1,10 +1,10 @@
 'use strict';
 
-module.exports.report = () => '❌ add react import';
+module.exports.report = () => '✅ add react import';
 
 module.exports.match = () => ({
     'import __imports from "__a"': ({__imports, __a}) => {
-        return __a.value !== 'react' || !isImport(__imports[0], 'ImportDefaultSpecifier', 'React');
+        return __a.value === 'react' && isImport(__imports[0], 'ImportDefaultSpecifier', 'React');
     },
 });
