+++ lib/plugin/index.js
@@ -8,26 +8,9 @@ const t = {
     'task-list-accept-task-prop': 'export const TaskList = (__args) => __body',
 };
 
-const isImport = (__import, type, name) => {
-    if (!__import)
-        return false;
-    
-    return __import.type === type && __import.local.name === name;
-};
-
 module.exports.rules = {
     ...getRule('-react-import'),
-    '+import-react': {
-        report: () => '✅ add react import',
-        match: () => ({
-            'import __imports from "__a"': ({__imports, __a}) => {
-                return __a.value === 'react' && isImport(__imports[0], 'ImportDefaultSpecifier', 'React');
-            },
-        }),
-        replace: () => ({
-            'import __imports from "__a"': ``,
-        }),
-    },
+    ...getRule('+react-import'),
     '-task-list-accept-task-prop': {
         report: () => '❌ &lt;TaskList> component should accept a tasks prop',
         match: () => ({
